requirements.txt
Flask==2.0.3
pytesseract==0.3.9
Pillow==8.3.2
